requirements.txt
bokeh
numpy
opencv-python
Pillow
pandas
ipywidgets